Now data of team ranking area on home page is taken from database, changed Id type from Long to Integer

+++ src/main/java/cstv/Interfaces/RoleRepository.java
@@ -3,7 +3,7 @@ package cstv.Interfaces;
 import cstv.Models.Role;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface RoleRepository extends MongoRepository<Role, Long> {
+public interface RoleRepository extends MongoRepository<Role, Integer> {
 
     Role findByRole(String role);
 }
