+++ src/main/java/cstv/Interfaces/TeamRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface TeamRepository extends MongoRepository<Team, Long> {
+public interface TeamRepository extends MongoRepository<Team, Integer> {
 
     Team findByName(String name);
